Do not track package-lock
This is controversial. Going with the side of staging the lock file
if it's an app, and ignoring it if it's a package.

This could make it a bit more difficult for contributors to start on the
project because the dependencies they download could be a newer version
that the maintainers are on. If this turns out to be a common issue, we
can always add the lock file back in.