Document via test why --help does not work
Because we use the git launcher to be able to provide subscripts to git
(such as `git mob`), we're also restricted by what the launcher allows.

Git expects a manual page when it sees --help, but it's out of scope for
us to provide a manpage for git-mob at this stage.