:sparkles: Show current mob
Will extend the test with more co-authors soon, should already work out
of the box.

I'm also thinking of including the primary author in the output, so it
gives the user a better idea of who the people on the final commit will
be. If we decide to do that, this could also change the behaviour for
the case where a mob isn't set -- do we show nothing (current behaviour)
or always at least show the primary author?