Clean up empty git config section after git-mob tests
Default Git behaviour is to keep empty sections in config even when
their last item has been removed. But like a good scout, we leave
the environment in a clean state when we're done.

Note the 'after()' call. This is to register the hook to always run,
regardless of whether there were earlier failures or if all tests were
skipped. Which is the behaviour we want for cleanup tasks.