:memo: Update package author to be more specific
Findmypast OSS is the name of the GitHub organisation we'll be using.
But it's more accurate to use the name of the company, aka our employer.
Even though most of the development is happening in our personal time
and with our personal resources, we also work on this with some of the
personal development time and resources that we get from work.