:sparkles: moved .git-coauthors, json format of co-authors.
Following the json format in issue #9. User initials for coauthor key which has name and emails as properties.

Removed the use of js-yaml and now read directly from a new file called .git-coauthors. Expected to be in json format.

Updated tests to reflect this change. Doc will need to be updated when ready to release.