:fire: Remove duplicate git-mob test
We already have two tests for checking that git-mob writes to
a .gitmessage file, for scenarios where the file already exists and when
it doesn't.

The extra test was redundant, and wasn't very good as it depended on
state from previous tests. According to its description it was testing
the scenario when the .gitmessage file doesn't exist, but it wasn't
deleting it in the test setup so it could exist from previous tests that
didn't perform cleanup.