:art: Move beforeEach state reset to afterEach hook
We were cleaning up the same stuff twice. Just do it once after every
test, but ensure that the cleanup always happens in case the test fails.
(That's probably why we had some cleanup happening before, in case the
previous test fails. But we can use the always() hook for that.)