Add gitRepo test helper for creating temporary isolated git repositories
So far we've been taking advantage of the project being in a git repo
itself, so we're running tests in our development environment.

This isn't ideal because our test environment can interfere with our
development environment's global state. Tests can alter the dev
environment and leave it in a damaged state. It can also cause tests to
be flaky when making changes to the dev environment during a test run.

Another benefit of having an isolated environment is that we can start
having true integration tests, that is, creating commits and seeing the
final result, without polluting the dev environment (even if it's just
the reflog). This gives us more confidence that the full flow works.