First step to solving issue #22 inline messages
Plan is to allow for a customisable approach to setup a
prepare-commit-msg hook to enable inline message.

I think it would be risk for git-mob to install the hook into project
which might already them setup and potentially breaking it.

Fallback will still be using the standard git template with the option
to use the prepare hook.