Don't append co-authors to commit message more than once
Let's make sure that the output of `git mob-print` isn't already present
in the commit message before we append it.

This avoids an issue whereby amending a commit or interactive rebasing
could lead to the co-authors being added to the commit message
multiple times.

It would probably be even better to match each coauthor individually,
since coauthors could still be added multiple times if the mob changes,
but that's a bigger change and in the mean time this still seems like
an improvement.

PR: https://github.com/findmypast-oss/git-mob/pull/52